# truck-service-host/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Truck Service API Documentation",
        "description": "Interactive API documentation for the Truck Service application. Developed by Mohammed Sleem.",
        "contact": {
            "email": "dev.mohammed@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://bull-station.com",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/my-trucks": {
            "get": {
                "tags": [
                    "My Trucks"
                ],
                "summary": "Get the authenticated user's trucks",
                "description": "Returns a paginated list of the user's own trucks, with filtering options.",
                "operationId": "getMyTrucks",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "latest",
                                "oldest"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MyTruckResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MyTruckResource": {
                "title": "My Truck Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "model": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "active",
                            "inactive"
                        ]
                    },
                    "price_per_day": {
                        "type": "string",
                        "format": "decimal"
                    },
                    "category": {
                        "type": "string"
                    },
                    "main_image": {
                        "type": "string",
                        "format": "url"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}